Refactor frontend study session into clean architecture layers
