+++ frontend/src/components/StudySession.tsx
@@ -1,25 +1,8 @@
 "use client";
 
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { fetchQuizzes, saveStudyResult } from "@/lib/api";
-import {
-  buildQuizSet,
-  calculateStudyResult,
-  countIncrementalMistakes,
-  getCharacterStates
-} from "@/lib/study";
-import {
-  appendMistakeLog,
-  appendReviewQueue,
-  appendStudyResult,
-  getReviewQueue,
-  getSettings,
-  removeRecoveredQuizIds,
-  saveLatestResult
-} from "@/lib/storage";
-import type { Quiz, QuizProgress, StudyMode } from "@/types/study";
+import { useStudySession } from "@/presentation/hooks/useStudySession";
+import type { StudyMode } from "@/types/study";
 
 function formatMs(ms: number) {
   const totalSeconds = Math.floor(ms / 1000);
@@ -31,99 +14,21 @@ function formatMs(ms: number) {
 }
 
 export function StudySession({ mode }: { mode: StudyMode }) {
-  const router = useRouter();
-  const [quizSet, setQuizSet] = useState<Quiz[]>([]);
-  const [isEmptyQuizSet, setIsEmptyQuizSet] = useState(false);
-  const [isReady, setIsReady] = useState(false);
-  const [loadError, setLoadError] = useState(false);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [inputValue, setInputValue] = useState("");
-  const [elapsedMs, setElapsedMs] = useState(0);
-  const [progressList, setProgressList] = useState<QuizProgress[]>([]);
-  const [quizStartedAt, setQuizStartedAt] = useState<number | null>(null);
-  const [wasMistaken, setWasMistaken] = useState(false);
-  const [mistakeCount, setMistakeCount] = useState(0);
-  const [isSavingResult, setIsSavingResult] = useState(false);
-
-  useEffect(() => {
-    const abortController = new AbortController();
-
-    setIsReady(false);
-    setLoadError(false);
-    setCurrentIndex(0);
-    setInputValue("");
-    setElapsedMs(0);
-    setProgressList([]);
-    setQuizStartedAt(null);
-    setWasMistaken(false);
-    setMistakeCount(0);
-    setIsSavingResult(false);
-
-    const loadQuizSet = async () => {
-      try {
-        const settings = getSettings();
-        const quizzes = await fetchQuizzes(
-          mode === "learn"
-            ? {
-                eikenLevels: settings.eikenLevels,
-                questionTypes: settings.questionTypes
-              }
-            : {},
-          abortController.signal
-        );
-        const reviewQueue = getReviewQueue();
-        const nextQuizSet = buildQuizSet(
-          quizzes,
-          mode,
-          reviewQueue,
-          settings.eikenLevels,
-          settings.questionTypes
-        );
-
-        setQuizSet(nextQuizSet);
-        setIsEmptyQuizSet(mode === "learn" && nextQuizSet.length === 0);
-        setIsReady(true);
-      } catch (error) {
-        if ((error as Error).name === "AbortError") {
-          return;
-        }
-
-        setQuizSet([]);
-        setIsEmptyQuizSet(false);
-        setLoadError(true);
-        setIsReady(true);
-      }
-    };
-
-    void loadQuizSet();
-
-    return () => {
-      abortController.abort();
-    };
-  }, [mode]);
-
-  const currentQuiz = quizSet[currentIndex];
-
-  useEffect(() => {
-    if (!currentQuiz || quizStartedAt !== null) {
-      return;
-    }
-
-    // クイズ切り替え時点を保持して平均入力時間を計測する。
-    setQuizStartedAt(Date.now());
-  }, [currentQuiz, quizStartedAt]);
-
-  useEffect(() => {
-    if (!currentQuiz) {
-      return;
-    }
-
-    const timerId = window.setInterval(() => {
-      setElapsedMs((value) => value + 100);
-    }, 100);
-
-    return () => window.clearInterval(timerId);
-  }, [currentQuiz]);
+  const {
+    characterStates,
+    currentIndex,
+    currentQuiz,
+    elapsedMs,
+    handleChange,
+    inputValue,
+    isEmptyQuizSet,
+    isReady,
+    isSavingResult,
+    loadError,
+    mistakeCount,
+    quizSet,
+    wasMistaken
+  } = useStudySession(mode);
 
   if (!isReady) {
     return (
@@ -183,95 +88,6 @@ export function StudySession({ mode }: { mode: StudyMode }) {
   if (!currentQuiz) {
     return null;
   }
-
-  const characterStates = getCharacterStates(currentQuiz.english, inputValue);
-
-  const finishQuiz = () => {
-    if (quizStartedAt === null) {
-      return;
-    }
-
-    const completedAt = new Date().toISOString();
-    const nextProgress: QuizProgress = {
-      quizId: currentQuiz.id,
-      durationMs: Date.now() - quizStartedAt,
-      mistakeCount,
-      wasMistaken,
-      completedAt
-    };
-    const nextProgressList = [...progressList, nextProgress];
-    const nextMistakeIds = nextProgressList
-      .filter((progress) => progress.wasMistaken)
-      .map((progress) => progress.quizId);
-    const recoveredIds = nextProgressList
-      .filter((progress) => !progress.wasMistaken)
-      .map((progress) => progress.quizId);
-
-    setProgressList(nextProgressList);
-
-    if (currentIndex === quizSet.length - 1) {
-      if (nextMistakeIds.length > 0) {
-        appendReviewQueue(nextMistakeIds);
-        appendMistakeLog(
-          nextProgressList
-            .filter((progress) => progress.wasMistaken)
-            .map((progress) => ({
-              question_id: progress.quizId,
-              mistake_count: progress.mistakeCount,
-              created_at: progress.completedAt
-            }))
-        );
-      }
-
-      if (mode === "review" && recoveredIds.length > 0) {
-        removeRecoveredQuizIds(recoveredIds);
-      }
-
-      const summary = calculateStudyResult(nextProgressList, mode);
-      saveLatestResult(summary);
-      appendStudyResult(summary);
-      setIsSavingResult(true);
-      void saveStudyResult(summary)
-        .catch(() => undefined)
-        .finally(() => {
-          router.push("/result");
-        });
-      return;
-    }
-
-    setCurrentIndex((value) => value + 1);
-    setInputValue("");
-    setQuizStartedAt(null);
-    setWasMistaken(false);
-    setMistakeCount(0);
-  };
-
-  const handleChange = (nextValue: string) => {
-    if (isSavingResult) {
-      return;
-    }
-
-    if (nextValue.length > currentQuiz.english.length) {
-      return;
-    }
-
-    const nextMistakes = countIncrementalMistakes(
-      inputValue,
-      nextValue,
-      currentQuiz.english
-    );
-
-    setInputValue(nextValue);
-
-    if (nextMistakes > 0) {
-      setWasMistaken(true);
-      setMistakeCount((value) => value + nextMistakes);
-    }
-
-    if (nextValue === currentQuiz.english) {
-      finishQuiz();
-    }
-  };
 
   return (
     <main className="page-shell">
