+++ frontend/src/domain/services/studyService.ts
@@ -1,0 +1,105 @@
+import type {
+  EikenLevel,
+  Quiz,
+  QuizProgress,
+  QuizType,
+  StudyMode,
+  StudyResult
+} from "@/types/study";
+
+export type CharacterState = "correct" | "wrong" | "pending";
+export const SESSION_QUESTION_COUNT = 10;
+
+function shuffleQuizzes(items: Quiz[]): Quiz[] {
+  const shuffled = [...items];
+
+  for (let index = shuffled.length - 1; index > 0; index -= 1) {
+    const swapIndex = Math.floor(Math.random() * (index + 1));
+    [shuffled[index], shuffled[swapIndex]] = [shuffled[swapIndex], shuffled[index]];
+  }
+
+  return shuffled;
+}
+
+export function buildQuizSet(
+  quizzes: Quiz[],
+  mode: StudyMode,
+  missedIds: number[],
+  eikenLevels: EikenLevel[] = ["5", "4", "3", "pre2", "2", "pre1", "1"],
+  questionTypes: QuizType[] = ["word", "sentence"],
+  sessionQuestionCount = SESSION_QUESTION_COUNT
+): Quiz[] {
+  if (mode === "review") {
+    const uniqueMissedIds = Array.from(new Set(missedIds));
+    const reviewSet = quizzes.filter((quiz) => uniqueMissedIds.includes(quiz.id));
+    return shuffleQuizzes(reviewSet).slice(0, sessionQuestionCount);
+  }
+
+  const filtered = quizzes.filter(
+    (quiz) =>
+      (eikenLevels.length === 0 || eikenLevels.includes(quiz.eikenLevel)) &&
+      (questionTypes.length === 0 || questionTypes.includes(quiz.type))
+  );
+  return shuffleQuizzes(filtered).slice(0, Math.min(sessionQuestionCount, filtered.length));
+}
+
+export function getCharacterStates(target: string, input: string): CharacterState[] {
+  return Array.from(target).map((character, index) => {
+    const typed = input[index];
+
+    if (typed === undefined) {
+      return "pending";
+    }
+
+    return typed === character ? "correct" : "wrong";
+  });
+}
+
+export function countIncrementalMistakes(
+  previousInput: string,
+  nextInput: string,
+  target: string
+): number {
+  if (!nextInput.startsWith(previousInput) || nextInput.length <= previousInput.length) {
+    return 0;
+  }
+
+  let nextMistakes = 0;
+
+  for (let index = previousInput.length; index < nextInput.length; index += 1) {
+    if (nextInput[index] !== target[index]) {
+      nextMistakes += 1;
+    }
+  }
+
+  return nextMistakes;
+}
+
+export function calculateStudyResult(
+  progressList: QuizProgress[],
+  mode: StudyMode
+): StudyResult {
+  const totalProblems = progressList.length;
+  const totalMistakes = progressList.reduce(
+    (accumulator, progress) => accumulator + progress.mistakeCount,
+    0
+  );
+  const perfectAnswers = progressList.filter((progress) => !progress.wasMistaken).length;
+  const totalDurationMs = progressList.reduce(
+    (accumulator, progress) => accumulator + progress.durationMs,
+    0
+  );
+
+  return {
+    mode,
+    total_questions: totalProblems,
+    correct_rate:
+      totalProblems === 0
+        ? 0
+        : Math.round((perfectAnswers / totalProblems) * 100),
+    mistakes: totalMistakes,
+    average_time:
+      totalProblems === 0 ? 0 : Math.round(totalDurationMs / totalProblems),
+    created_at: new Date().toISOString()
+  };
+}
